fix tests

+++ src/test/java/jcifs/smb/NtStatusTest.java
@@ -18,10 +18,10 @@ class NtStatusTest {
     void testWellKnownStatusConstants() {
         // Verify important NT status constants are defined
         assertEquals(0x00000000, NtStatus.NT_STATUS_SUCCESS);
-        assertEquals(0xC0000022L, NtStatus.NT_STATUS_ACCESS_DENIED);
-        assertEquals(0xC0000034L, NtStatus.NT_STATUS_OBJECT_NAME_NOT_FOUND);
-        assertEquals(0xC0000043L, NtStatus.NT_STATUS_SHARING_VIOLATION);
-        assertEquals(0xC000000DL, NtStatus.NT_STATUS_INVALID_PARAMETER);
+        assertEquals((int)0xC0000022L, NtStatus.NT_STATUS_ACCESS_DENIED);
+        assertEquals((int)0xC0000034L, NtStatus.NT_STATUS_OBJECT_NAME_NOT_FOUND);
+        assertEquals((int)0xC0000043L, NtStatus.NT_STATUS_SHARING_VIOLATION);
+        assertEquals((int)0xC000000DL, NtStatus.NT_STATUS_INVALID_PARAMETER);
     }
 
     @Test
@@ -48,7 +48,7 @@ class NtStatusTest {
 
         // When/Then - All error codes should have high bit set (0xC prefix)
         for (int status : errorStatuses) {
-            assertTrue((status & 0xC0000000L) == 0xC0000000L, 
+            assertTrue((status & (int)0xC0000000L) == (int)0xC0000000L, 
                 "Error status should have severity bits set to 0xC: " + Integer.toHexString(status));
         }
     }
@@ -91,9 +91,9 @@ class NtStatusTest {
         int status = NtStatus.NT_STATUS_ACCESS_DENIED;
 
         // When/Then
-        assertEquals(0xC0000022L, status);
+        assertEquals((int)0xC0000022L, status);
         // Should be an error status (0xC prefix)
-        assertTrue((status & 0xC0000000L) == 0xC0000000L);
+        assertTrue((status & (int)0xC0000000L) == (int)0xC0000000L);
     }
 
     @Test
@@ -103,9 +103,9 @@ class NtStatusTest {
         int status = NtStatus.NT_STATUS_OBJECT_NAME_NOT_FOUND;
 
         // When/Then
-        assertEquals(0xC0000034L, status);
+        assertEquals((int)0xC0000034L, status);
         // Should be an error status (0xC prefix)
-        assertTrue((status & 0xC0000000L) == 0xC0000000L);
+        assertTrue((status & (int)0xC0000000L) == (int)0xC0000000L);
     }
 
     @Test
@@ -115,9 +115,9 @@ class NtStatusTest {
         int status = NtStatus.NT_STATUS_SHARING_VIOLATION;
 
         // When/Then
-        assertEquals(0xC0000043L, status);
+        assertEquals((int)0xC0000043L, status);
         // Should be an error status (0xC prefix)
-        assertTrue((status & 0xC0000000L) == 0xC0000000L);
+        assertTrue((status & (int)0xC0000000L) == (int)0xC0000000L);
     }
 
     @Test
@@ -127,9 +127,9 @@ class NtStatusTest {
         int status = NtStatus.NT_STATUS_LOGON_FAILURE;
 
         // When/Then
-        assertEquals(0xC000006DL, status);
+        assertEquals((int)0xC000006DL, status);
         // Should be an error status (0xC prefix)
-        assertTrue((status & 0xC0000000L) == 0xC0000000L);
+        assertTrue((status & (int)0xC0000000L) == (int)0xC0000000L);
     }
 
     @Test
@@ -139,9 +139,9 @@ class NtStatusTest {
         int infoStatus = NtStatus.NT_STATUS_BUFFER_OVERFLOW;
 
         // When/Then
-        assertEquals(0x80000005L, infoStatus);
+        assertEquals((int)0x80000005L, infoStatus);
         // Should be an informational/warning status (0x8 prefix)
-        assertTrue((infoStatus & 0x80000000L) == 0x80000000L);
+        assertTrue((infoStatus & (int)0x80000000L) == (int)0x80000000L);
     }
 
     @Test
@@ -151,9 +151,9 @@ class NtStatusTest {
         int pendingStatus = NtStatus.NT_STATUS_PENDING;
 
         // When/Then
-        assertEquals(0x00000103L, pendingStatus);
+        assertEquals((int)0x00000103L, pendingStatus);
         // Pending status is informational (not error)
-        assertTrue((pendingStatus & 0xC0000000L) != 0xC0000000L);
+        assertTrue((pendingStatus & (int)0xC0000000L) != (int)0xC0000000L);
     }
 
     @Test
